fixing admin page
