+++ apps/console/src/components/command-palette.tsx
@@ -20,6 +20,9 @@ import { useRouter } from "next/navigation"
 import { useCallback, useEffect, useRef, useState } from "react"
 import { createPortal } from "react-dom"
 
+import { useUser } from "@/hooks/use-user"
+import { canReadPlatformTemplates } from "@/lib/admin/permissions"
+
 interface CommandItem {
   label: string
   icon: Icon
@@ -30,6 +33,7 @@ export function CommandPalette() {
   const [open, setOpen] = useState(false)
   const [hoveredLabel, setHoveredLabel] = useState<string | null>(null)
   const router = useRouter()
+  const { user } = useUser()
 
   useEffect(() => {
     const handleKeyDown = (e: KeyboardEvent) => {
@@ -124,6 +128,13 @@ export function CommandPalette() {
 
   const inputRef = useRef<HTMLInputElement>(null)
   const [mounted, setMounted] = useState(false)
+  const filteredNavigationItems = navigationItems.filter(
+    (item) => item.label !== "Templates" || canReadPlatformTemplates(user),
+  )
+  const filteredActionItems = actionItems.filter(
+    (item) =>
+      item.label !== "Create template" || canReadPlatformTemplates(user),
+  )
 
   useEffect(() => {
     setMounted(true)
@@ -189,7 +200,7 @@ export function CommandPalette() {
                   heading="Navigation"
                   className="[&_[cmdk-group-heading]]:px-2.5 [&_[cmdk-group-heading]]:py-1.5 [&_[cmdk-group-heading]]:font-mono [&_[cmdk-group-heading]]:text-xs [&_[cmdk-group-heading]]:text-muted [&_[cmdk-group-heading]]:uppercase"
                 >
-                  {navigationItems.map((item) => (
+                  {filteredNavigationItems.map((item) => (
                     <Command.Item
                       key={item.label}
                       onSelect={item.onSelect}
@@ -222,7 +233,7 @@ export function CommandPalette() {
                   heading="Actions"
                   className="[&_[cmdk-group-heading]]:px-2.5 [&_[cmdk-group-heading]]:py-1.5 [&_[cmdk-group-heading]]:font-mono [&_[cmdk-group-heading]]:text-xs [&_[cmdk-group-heading]]:text-muted [&_[cmdk-group-heading]]:uppercase"
                 >
-                  {actionItems.map((item) => (
+                  {filteredActionItems.map((item) => (
                     <Command.Item
                       key={item.label}
                       onSelect={item.onSelect}
