+++ apps/console/src/lib/api/proxy-auth.ts
@@ -8,6 +8,7 @@ import {
 } from "@/lib/admin/impersonation"
 import { platformImpersonationReadScopes } from "@/lib/admin/permissions"
 import { ensureImpersonationKeyRow } from "@/lib/admin/impersonation-key"
+import { platformImpersonationReadScopes } from "@/lib/admin/permissions"
 import { getProxySecret, hashKey } from "@/lib/api/proxy-secret"
 import {
   pickActiveTeam,
